fix: doc and example
