+++ docs/enums/RequestStatus.html
@@ -19,7 +19,7 @@
 <div class="tsd-comment tsd-typography"><p>The status of the Data Subject Request.</p>
 </div></section><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/df04f00/src/index.ts#L25">index.ts:25</a></li></ul></aside>
+<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/6231af4/src/index.ts#L25">index.ts:25</a></li></ul></aside>
 <section class="tsd-panel-group tsd-index-group">
 <section class="tsd-panel tsd-index-panel">
 <details class="tsd-index-content tsd-index-accordion" open><summary class="tsd-accordion-summary tsd-index-summary">
@@ -40,32 +40,32 @@
 <h3 class="tsd-anchor-link"><span>Cancelled<wbr/>Request<wbr/>Status</span><a href="#CancelledRequestStatus" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><path stroke="none" d="M0 0h24v24H0z" fill="none" id="icon-anchor-a"></path><path d="M10 14a3.5 3.5 0 0 0 5 0l4 -4a3.5 3.5 0 0 0 -5 -5l-.5 .5" id="icon-anchor-b"></path><path d="M14 10a3.5 3.5 0 0 0 -5 0l-4 4a3.5 3.5 0 0 0 5 5l.5 -.5" id="icon-anchor-c"></path></svg></a></h3>
 <div class="tsd-signature"><span class="tsd-kind-enum-member">Cancelled<wbr/>Request<wbr/>Status</span><span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">&quot;cancelled&quot;</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/df04f00/src/index.ts#L30">index.ts:30</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/6231af4/src/index.ts#L30">index.ts:30</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member"><a id="CompletedRequestStatus" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><span>Completed<wbr/>Request<wbr/>Status</span><a href="#CompletedRequestStatus" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature"><span class="tsd-kind-enum-member">Completed<wbr/>Request<wbr/>Status</span><span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">&quot;completed&quot;</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/df04f00/src/index.ts#L29">index.ts:29</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/6231af4/src/index.ts#L29">index.ts:29</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member"><a id="DeniedRequestStatus" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><span>Denied<wbr/>Request<wbr/>Status</span><a href="#DeniedRequestStatus" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature"><span class="tsd-kind-enum-member">Denied<wbr/>Request<wbr/>Status</span><span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">&quot;denied&quot;</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/df04f00/src/index.ts#L31">index.ts:31</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/6231af4/src/index.ts#L31">index.ts:31</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member"><a id="InProgressRequestStatus" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><span>In<wbr/>Progress<wbr/>Request<wbr/>Status</span><a href="#InProgressRequestStatus" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature"><span class="tsd-kind-enum-member">In<wbr/>Progress<wbr/>Request<wbr/>Status</span><span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">&quot;in_progress&quot;</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/df04f00/src/index.ts#L28">index.ts:28</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/6231af4/src/index.ts#L28">index.ts:28</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member"><a id="PendingRequestStatus" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><span>Pending<wbr/>Request<wbr/>Status</span><a href="#PendingRequestStatus" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature"><span class="tsd-kind-enum-member">Pending<wbr/>Request<wbr/>Status</span><span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">&quot;pending&quot;</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/df04f00/src/index.ts#L27">index.ts:27</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/6231af4/src/index.ts#L27">index.ts:27</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member"><a id="UnknownRequestStatus" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><span>Unknown<wbr/>Request<wbr/>Status</span><a href="#UnknownRequestStatus" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature"><span class="tsd-kind-enum-member">Unknown<wbr/>Request<wbr/>Status</span><span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">&quot;unknown&quot;</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/df04f00/src/index.ts#L26">index.ts:26</a></li></ul></aside></section></section></div>
+<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/6231af4/src/index.ts#L26">index.ts:26</a></li></ul></aside></section></section></div>
 <div class="col-sidebar">
 <div class="page-menu">
 <div class="tsd-navigation settings">
