+++ docs/index.html
@@ -17,7 +17,7 @@
 When the transponder is connected and authenticated to Ketch, the agent periodically checks for any DSR Right activity that needs to be executed.
 When a user invokes their DSR Right, the right is forwarded to the agent which then downloads your nodejs module and executes it dynamically.
 Currently, we support scripts located in your local file system, or hosted on a public http server, or in an AWS S3 bucket behind default S3 authentication.</p>
-<a id="md:entrypoint" class="tsd-anchor"></a><h1><a href="#$entrypoint" style="color:inherit;text-decoration:none">Entrypoint</a></h1><p>Your module need to export the <a href="./types/HandleRequest.html">HandleRequest</a> function as defined in this documentation. You can use handlers exposed as part of the argument to this function to interact with the Ketch platform at runtime.</p>
+<a id="md:entrypoint" class="tsd-anchor"></a><h1><a href="#$entrypoint" style="color:inherit;text-decoration:none">Entrypoint</a></h1><p>Your module need to export the <a href="./types/HandleRequest.html">HandleRequest</a> function as defined in <a href="https://ketch-sdk.github.io/ketch-agent/index.html">the documentation</a>. You can use handlers exposed as part of the argument to this function to interact with the Ketch platform at runtime.</p>
 <a id="md:runtime-of-ketch-agent" class="tsd-anchor"></a><h1><a href="#$runtime-of-ketch-agent" style="color:inherit;text-decoration:none">Runtime of Ketch agent</a></h1><ul>
 <li>Nodejs LTS/hydrogen (v18.16.0)</li>
 </ul>
