+++ docs/types/DSRVersion.html
@@ -17,7 +17,7 @@
 <h1>Type alias DSRVersion</h1></div>
 <div class="tsd-signature"><span class="tsd-kind-type-alias">DSRVersion</span><span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">&quot;dsr/v1&quot;</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/df04f00/src/index.ts#L20">index.ts:20</a></li></ul></aside></div>
+<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/6231af4/src/index.ts#L20">index.ts:20</a></li></ul></aside></div>
 <div class="col-sidebar">
 <div class="page-menu">
 <div class="tsd-navigation settings">
