+++ docs/types/DsrRequest.html
@@ -19,7 +19,7 @@
 <div class="tsd-comment tsd-typography"><p>DSR Request</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/df04f00/src/index.ts#L207">index.ts:207</a></li></ul></aside></div>
+<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/6231af4/src/index.ts#L207">index.ts:207</a></li></ul></aside></div>
 <div class="col-sidebar">
 <div class="page-menu">
 <div class="tsd-navigation settings">
