+++ docs/types/DsrResponse.html
@@ -19,7 +19,7 @@
 <div class="tsd-comment tsd-typography"><p>DSR Response</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/df04f00/src/index.ts#L212">index.ts:212</a></li></ul></aside></div>
+<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/6231af4/src/index.ts#L212">index.ts:212</a></li></ul></aside></div>
 <div class="col-sidebar">
 <div class="page-menu">
 <div class="tsd-navigation settings">
