+++ docs/types/HandleRequest.html
@@ -34,7 +34,7 @@
 <h5><span class="tsd-kind-parameter">connectionConfig</span>: <a href="../interfaces/ConnectionConfig.html" class="tsd-signature-type tsd-kind-interface">ConnectionConfig</a></h5></li></ul></div>
 <h4 class="tsd-returns-title">Returns <span class="tsd-signature-type ">Promise</span><span class="tsd-signature-symbol">&lt;</span><a href="DSRResponse.html" class="tsd-signature-type tsd-kind-type-alias">DSRResponse</a><span class="tsd-signature-symbol">&gt;</span></h4></li></ul></li></ul></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/df04f00/src/index.ts#L18">index.ts:18</a></li></ul></aside></div>
+<li>Defined in <a href="https://github.com/ketch-sdk/ketch-agent/blob/6231af4/src/index.ts#L18">index.ts:18</a></li></ul></aside></div>
 <div class="col-sidebar">
 <div class="page-menu">
 <div class="tsd-navigation settings">
