--- examples/mongodsr/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "mongodsr",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "mongodb": "^5.4.0",
-    "typescript": "^5.0.4"
-  },
-  "devDependencies": {
-    "@ketch-sdk/ketch-agent": "^0.0.1"
-  }
-}
