+++ package.json
@@ -46,6 +46,7 @@
     "typescript-formatter": "^7.2.2"
   },
   "dependencies": {
+    "mongodb": "^5.7.0",
     "tslib": "^2.5.0"
   }
 }
